+++ src/contexts/NotificationContext.jsx
@@ -1,0 +1,34 @@
+import { createContext, useState, useContext } from 'react';
+import { NotificationContainer } from '../components/Notifications';
+
+const NotificationContext = createContext();
+
+export const NotificationProvider = ({ children }) => {
+  const [notifications, setNotifications] = useState([]);
+
+  const addNotification = (type, message, duration = 3000) => {
+    const id = Date.now();
+    const notification = { id, type, message };
+
+    setNotifications(prev => [...prev, notification]);
+
+    setTimeout(() => {
+      setNotifications(prev => prev.filter(n => n.id !== id));
+    }, duration);
+  };
+
+  return (
+    <NotificationContext.Provider value={{ addNotification }}>
+      {children}
+      <NotificationContainer notifications={notifications} />
+    </NotificationContext.Provider>
+  );
+};
+
+export const useNotification = () => {
+  const context = useContext(NotificationContext);
+  if (!context) {
+    throw new Error('useNotification must be used within NotificationProvider');
+  }
+  return context;
+};
