+++ src/services/websocketService.js
@@ -1,0 +1,84 @@
+// ==================== WEBSOCKET SERVICE ====================
+class WebSocketService {
+  constructor() {
+    this.ws = null;
+    this.listeners = new Map();
+    this.reconnectInterval = 5000;
+    this.shouldReconnect = true;
+  }
+
+  connect() {
+    if (!CONFIG.ENABLE_WEBSOCKETS) return;
+
+    try {
+      this.ws = new WebSocket(CONFIG.WEBSOCKET_URL);
+
+      this.ws.onopen = () => {
+        console.log('WebSocket conectado');
+        this.notifyListeners('connection', { status: 'connected' });
+      };
+
+      this.ws.onmessage = (event) => {
+        try {
+          const data = JSON.parse(event.data);
+          this.notifyListeners(data.type, data.payload);
+        } catch (error) {
+          console.error('Error parsing WebSocket message:', error);
+        }
+      };
+
+      this.ws.onclose = () => {
+        console.log('WebSocket desconectado');
+        this.notifyListeners('connection', { status: 'disconnected' });
+
+        if (this.shouldReconnect) {
+          setTimeout(() => this.connect(), this.reconnectInterval);
+        }
+      };
+
+      this.ws.onerror = (error) => {
+        console.error('WebSocket error:', error);
+      };
+    } catch (error) {
+      console.error('Error conectando WebSocket:', error);
+    }
+  }
+
+  disconnect() {
+    this.shouldReconnect = false;
+    if (this.ws) {
+      this.ws.close();
+    }
+  }
+
+  subscribe(eventType, callback) {
+    if (!this.listeners.has(eventType)) {
+      this.listeners.set(eventType, new Set());
+    }
+    this.listeners.get(eventType).add(callback);
+
+    // Retornar función para desuscribirse
+    return () => {
+      const callbacks = this.listeners.get(eventType);
+      if (callbacks) {
+        callbacks.delete(callback);
+      }
+    };
+  }
+
+  notifyListeners(eventType, data) {
+    const callbacks = this.listeners.get(eventType);
+    if (callbacks) {
+      callbacks.forEach(callback => callback(data));
+    }
+  }
+
+  send(type, payload) {
+    if (this.ws && this.ws.readyState === WebSocket.OPEN) {
+      this.ws.send(JSON.stringify({ type, payload }));
+    }
+  }
+}
+
+const wsService = new WebSocketService();
+export default wsService;
